GEODE-9786: Make IndexibleTreeSet inner class of quick check test
Authored-by: Donal Evans <doevans@vmware.com>
